# ui/src/generated/ponti-ai.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ponti AI",
    "version": "1.0.0-mvp"
  },
  "paths": {
...
    "/v1/insights/compute": {
      "post": {
        "summary": "Compute Insights",
        "operationId": "compute_insights_v1_insights_compute_post",
        "parameters": [
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeInsightsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insights/{entity_type}/{entity_id}": {
      "get": {
        "summary": "Get Insights",
        "operationId": "get_insights_v1_insights__entity_type___entity_id__get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insights/summary": {
      "get": {
        "summary": "Get Summary",
        "operationId": "get_summary_v1_insights_summary_get",
        "parameters": [
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insights/{insight_id}/actions": {
      "post": {
        "summary": "Record Action",
        "operationId": "record_action_v1_insights__insight_id__actions_post",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Insight Id"
            }
          },
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/copilot/insights/{insight_id}/explain": {
      "get": {
        "summary": "Copilot Explain",
        "operationId": "copilot_explain_v1_copilot_insights__insight_id__explain_get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Insight Id"
            }
          },
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainInsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/copilot/insights/{insight_id}/why": {
      "get": {
        "summary": "Copilot Why",
        "operationId": "copilot_why_v1_copilot_insights__insight_id__why_get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Insight Id"
            }
          },
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainInsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/copilot/insights/{insight_id}/next-steps": {
      "get": {
        "summary": "Copilot Next Steps",
        "operationId": "copilot_next_steps_v1_copilot_insights__insight_id__next_steps_get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Insight Id"
            }
          },
          {
            "name": "X-SERVICE-KEY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Key"
            }
          },
          {
            "name": "X-USER-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-PROJECT-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainInsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "minLength": 1,
            "title": "Action"
          },
          "new_status": {
            "type": "string",
            "minLength": 1,
            "title": "New Status"
          }
        },
        "type": "object",
        "required": [
          "action",
          "new_status"
        ],
        "title": "ActionRequest"
      },
      "ActionResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "service_kind": {
            "type": "string",
            "const": "insight_service",
            "title": "Service Kind",
            "default": "insight_service"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status"
        ],
        "title": "ActionResponse"
      },
      "ComputeInsightsResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "service_kind": {
            "type": "string",
            "const": "insight_service",
            "title": "Service Kind",
            "default": "insight_service"
          },
          "computed": {
            "type": "integer",
            "title": "Computed"
          },
          "insights_created": {
            "type": "integer",
            "title": "Insights Created"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "computed",
          "insights_created"
        ],
        "title": "ComputeInsightsResponse"
      },
      "CopilotExplanation": {
        "properties": {
          "human_readable": {
            "type": "string",
            "title": "Human Readable"
          },
          "audit_focused": {
            "type": "string",
            "title": "Audit Focused"
          },
          "what_to_watch_next": {
            "type": "string",
            "title": "What To Watch Next"
          }
        },
        "type": "object",
        "required": [
          "human_readable",
          "audit_focused",
          "what_to_watch_next"
        ],
        "title": "CopilotExplanation"
      },
      "ExplainInsightResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "output_kind": {
            "type": "string",
            "const": "copilot_explanation",
            "title": "Output Kind",
            "default": "copilot_explanation"
          },
          "routed_agent": {
            "type": "string",
            "const": "copilot",
            "title": "Routed Agent",
            "default": "copilot"
          },
          "routing_source": {
            "type": "string",
            "const": "copilot_agent",
            "title": "Routing Source",
            "default": "copilot_agent"
          },
          "insight_id": {
            "type": "string",
            "title": "Insight Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "explain",
              "why",
              "next-steps"
            ],
            "title": "Mode"
          },
          "explanation": {
            "$ref": "#/components/schemas/CopilotExplanation"
          },
          "proposal": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "insight_id",
          "mode",
          "explanation"
        ],
        "title": "ExplainInsightResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InsightItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "integer",
            "title": "Severity"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "explanations": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanations"
          },
          "action": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "features_version": {
            "type": "string",
            "title": "Features Version"
          },
          "computed_at": {
            "type": "string",
            "title": "Computed At"
          },
          "valid_until": {
            "type": "string",
            "title": "Valid Until"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "impact_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Min"
          },
          "impact_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Max"
          },
          "impact_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Unit"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          },
          "cooldown_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Until"
          },
          "computed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed By"
          },
          "job_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Run Id"
          },
          "rules_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "entity_type",
          "entity_id",
          "type",
          "severity",
          "priority",
          "title",
          "summary",
          "evidence",
          "explanations",
          "action",
          "model_version",
          "features_version",
          "computed_at",
          "valid_until",
          "status"
        ],
        "title": "InsightItem"
      },
      "InsightListResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "service_kind": {
            "type": "string",
            "const": "insight_service",
            "title": "Service Kind",
            "default": "insight_service"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/InsightItem"
            },
            "type": "array",
            "title": "Insights"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "insights"
        ],
        "title": "InsightListResponse"
      },
...
      "SummaryResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "service_kind": {
            "type": "string",
            "const": "insight_service",
            "title": "Service Kind",
            "default": "insight_service"
          },
          "output_kind": {
            "type": "string",
            "const": "insight_summary",
            "title": "Output Kind",
            "default": "insight_summary"
          },
          "new_count_total": {
            "type": "integer",
            "title": "New Count Total"
          },
          "new_count_high_severity": {
            "type": "integer",
            "title": "New Count High Severity"
          },
          "top_insights": {
            "items": {
              "$ref": "#/components/schemas/InsightItem"
            },
            "type": "array",
            "title": "Top Insights"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "new_count_total",
          "new_count_high_severity",
          "top_insights"
        ],
        "title": "SummaryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}